Move the mentor schema
