+++ src/types.ts
@@ -1,5 +1,8 @@
 import { type z } from 'zod';
-import { type MenteeApplicationSchema } from './schemas';
+import {
+  type MentorApplicationSchema,
+  type MenteeApplicationSchema,
+} from './schemas';
 
 export interface Category {
   category: string;
@@ -20,6 +23,8 @@ export interface Mentor {
 
 export type MenteeApplication = z.infer<typeof MenteeApplicationSchema>;
 
+export type MentorApplication = z.infer<typeof MentorApplicationSchema>;
+
 export interface Mentee {
   state: string;
   application: MenteeApplication;
@@ -30,46 +35,6 @@ export interface Mentee {
   updated_at: Date;
   certificate_id: string;
   journal: string;
-}
-
-// export interface MenteeApplication {
-//   firstName: string;
-//   lastName: string;
-//   email: string;
-//   contactNo: string;
-//   company?: string;
-//   course?: string;
-//   position: string;
-//   cv: string;
-//   isUndergrad: boolean;
-//   consentGiven: boolean;
-//   graduatedYear: number;
-//   mentorId: string;
-//   submission: string;
-// }
-
-export interface MentorApplication {
-  firstName: string;
-  lastName: string;
-  email: string;
-  contactNo: string;
-  country: string;
-  position: string;
-  expertise: string;
-  bio: string;
-  isPastMentor: boolean;
-  reasonToMentor?: string;
-  motivation?: string;
-  cv: string;
-  menteeExpectations: string;
-  mentoringPhilosophy: string;
-  noOfMentees: number;
-  canCommit: boolean;
-  mentoredYear?: number;
-  category: string;
-  institution: string;
-  linkedin?: string;
-  website?: string;
 }
 
 export interface Profile {
